画像管理機能のリファクタリングと共通ユーティリティの抽出
- S3関連の処理を共通のユーティリティ関数に抽出
- ファイルアップロード関連の処理を別モジュールに分離
- MinIO設定の取得と検証を共通化
- 画像URL生成ロジックを統一
- コードの重複を削減し、モジュール性を向上
